refactor: rename FolderSelectionDialog to follow kebab-case convention
- Rename FolderSelectionDialog.tsx to folder-selection-dialog.tsx per coding guidelines
- Update import path in SelectedCapsBar.tsx to use new kebab-case filename
- Preserve git history using git mv command
